pyproject.toml
[tool.poetry]
name = "treelyze"
version = "0.1.0"
description = "A CLI tool for analyzing directory structures and summarizing files"
authors = ["Magnus <magnus.uno.friberg@gmail.com>"]
readme = "README.md"
repository = "https://github.com/mandgie/treelyze"
keywords = ["cli", "directory", "analysis", "summarization"]

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
pyyaml = "^6.0.2"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[tool.poetry.scripts]
treelyze = "treelyze.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"